requirements.txt
gradio
tensorflow
numpy
Pillow
pandas
